fix derp

+++ services/core/jni/com_android_server_power_PowerManagerService.cpp
@@ -120,7 +120,7 @@ static void nativeAcquireSuspendBlocker(JNIEnv *env, jclass /* clazz */, jstring
 static void nativeCpuBoost(JNIEnv *env, jobject clazz, jint duration) {
     // Tell the Power HAL to boost the CPU
     if (gPowerModule && gPowerModule->powerHint) {
-        gPowerModule->powerHint(gPowerModule, POWER_HINT_CPU_BOOST, (void *) duration);
+        gPowerModule->powerHint(gPowerModule, POWER_HINT_CPU_BOOST, (void *)(intptr_t) duration);
     }
 }
